fix(loader, vmm, libc): fix ELF segment page alignment in loader, user mmap bounds, and squashed compiler warnings

+++ user/libc/src/printf.c
@@ -943,13 +943,13 @@ int vfctprintf(void (*out)(char character, void* arg), void* arg, const char* fo
 }
 
 void _fprintf_putchar(char c, void* arg) {
-    fputchar((int)arg, c);
+    fputchar((int)(uintptr_t)arg, c);
 }
 
 int fprintf(int fd, const char* fmt, ...) {
     va_list lst;
     va_start(lst, fmt);
-    const int ret = fctprintf(_fprintf_putchar, (void*)fd, fmt, lst);
+    const int ret = fctprintf(_fprintf_putchar, (void*)(uintptr_t)fd, fmt, lst);
     va_end(lst);
     return ret;
 }
